+++ R/rxode2.R
@@ -1286,7 +1286,7 @@ rxCompile <- function(model, dir, prefix, force = FALSE, modName = NULL,
       .cc <- gsub("\n", "", .cc)
       .cflags <- rawToChar(sys::exec_internal(file.path(R.home("bin"), "R"), c("CMD", "config", "CFLAGS"))$stdout)
       .cflags <- gsub("\n", "", .cflags)
-      .cflags <- paste0(.cflags, " -O", getOption("rxode2.compile.O", "2"))
+      .cflags <- paste0(.cflags, " -O", getOption("rxode2.compile.O", "3"))
       .shlibCflags <- rawToChar(sys::exec_internal(file.path(R.home("bin"), "R"), c("CMD", "config", "SHLIB_CFLAGS"))$stdout)
       .shlibCflags <- gsub("\n", "", .shlibCflags)
       .cpicflags <- rawToChar(sys::exec_internal(file.path(R.home("bin"), "R"), c("CMD", "config", "CPICFLAGS"))$stdout)
@@ -1533,8 +1533,8 @@ rxCompile.rxModelVars <- function(model, # Model
         }
         .defs <- ""
         .ret <- sprintf(
-          "#rxode2 Makevars\nPKG_CFLAGS=-O%s %s -I\"%s\" -I\"%s\"\nPKG_LIBS=$(BLAS_LIBS) $(LAPACK_LIBS) $(FLIBS)\n",
-          getOption("rxode2.compile.O", "2"),
+          "#rxode2 Makevars\nPKG_CFLAGS=-O%s -fno-math-errno %s -I\"%s\" -I\"%s\"\nPKG_LIBS=$(BLAS_LIBS) $(LAPACK_LIBS) $(FLIBS)\n",
+          getOption("rxode2.compile.O", "3"),
           .defs, .getIncludeDir(),
           system.file("include", package = "rxode2")
         )